Add files via upload
Heuristic Training data management is changed from the previous version. In 5.0, the user gives the agent a training model and immediately after the constructed training batch is processed for X number of epochs. In the newer version a heuristic module is added. The user now chooses between appending training data or use data from a file. This helps to save training samples and create a larger batch size.